Refactor: 네비게이션 관련 타입 수정에 의한 컴포넌트 코드 수정
